refactor searchPosting
