+++ src/main/java/com/board_of_ads/service/interfaces/PostingService.java
@@ -20,4 +20,6 @@ public interface PostingService {
     List<PostingDto> getPostingByFullRegionName(String name);
 
     List<PostingDto> getAllPostings();
+
+    List<PostingDto> searchPostings(String categorySelect, String citySelect, String searchText, String photoOption);
 }
