+++ src/main/resources/templates/fragments/search-header.html
@@ -3,7 +3,7 @@
     <ul id="menu-search">
 
     <li id="category">
-        <select id="category-select" class="categoriesSelect" onchange="getCategoryOption(this)">
+        <select id="category-select" class="categoriesSelect" onchange="getNewPostings()">
         </select>
     </li>
 
@@ -17,7 +17,7 @@
     </li>
 
         <li id="image-field">
-            <select id="image-select" name="изображение">
+            <select id="image-select" name="изображение" onchange="getNewPostings()">
                 <option th:text="#{allAds}" value="пункт1" selected="selected">Все объявления</option>
                 <option th:text="#{wFoto}" value="пункт2">С фото</option>
                 <option th:text="#{woFoto}" value="пункт3">Без фото</option>
@@ -34,7 +34,7 @@
 
     <li id="search-left"> <!--todo: link to action -->
         <button th:text="#{justsearch}" id="search-left-btn"
-                name="search-button" onclick="searchPosts()">
+                name="search-button" onclick="getNewPostings()">
             Найти
         </button>
     </li>
